requirements.txt
# DocMind MVP — зависимости
# Установка: pip install -r requirements.txt

# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6

# Документы
pymupdf>=1.23.0
python-docx>=1.1.0
fpdf2>=2.7.0

# AI
openai>=1.12.0

# Переменные окружения
python-dotenv>=1.0.0
